pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "insurance-ebm"
version = "0.1.0"
description = "Insurance workflow wrapper around interpretML's ExplainableBoostingMachine"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"
authors = [{ name = "Burning Cost", email = "pricing.frontier@gmail.com" }]
keywords = ["insurance", "pricing", "EBM", "explainable boosting", "interpretML", "actuarial"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "numpy >= 1.21",
    "polars >= 0.20",
    "matplotlib >= 3.4",
    "scikit-learn >= 1.0",
    "pandas>=2.3.3",
    "pyarrow>=21.0.0",
]

[project.optional-dependencies]
interpret = ["interpret >= 0.7.0"]
excel = ["openpyxl >= 3.0"]
glm = ["statsmodels >= 0.13"]
dev = [
    "pytest >= 7.0",
    "pytest-cov",
    "openpyxl >= 3.0",
    "statsmodels >= 0.13",
]
all = [
    "interpret >= 0.7.0",
    "openpyxl >= 3.0",
    "statsmodels >= 0.13",
]

[project.urls]
Homepage = "https://github.com/burning-cost/insurance-ebm"
Repository = "https://github.com/burning-cost/insurance-ebm"

[tool.hatch.build.targets.wheel]
packages = ["src/insurance_ebm"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"

[dependency-groups]
dev = [
    "databricks-sdk>=0.97.0",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
]
